+++ packages/pluggableWidgets/gallery-web/src/components/GalleryContent.tsx
@@ -2,14 +2,16 @@ import { useInfiniteControl } from "@mendix/widget-plugin-grid/components/Infini
 import classNames from "classnames";
 import { observer } from "mobx-react-lite";
 import { PropsWithChildren, ReactElement } from "react";
-import { usePaginationConfig, usePaginationVM } from "../model/hooks/injection-hooks";
+import { useLoaderViewModel, usePaginationConfig, usePaginationVM } from "../model/hooks/injection-hooks";
 
 export const GalleryContent = observer(function GalleryContent({ children }: PropsWithChildren): ReactElement {
     const paginationVM = usePaginationVM();
+    const loaderVM = useLoaderViewModel();
     const isInfinite = usePaginationConfig().isVirtualScrolling;
     const [trackScrolling, bodySize, containerRef] = useInfiniteControl({
         hasMoreItems: paginationVM.hasMoreItems,
         isInfinite,
+        isRefreshing: loaderVM.isRefreshing,
         setPage: paginationVM.setPage.bind(paginationVM)
     });